requirements.txt
bs4==0.0.1
discord.py==1.3.4
python-dotenv==0.14.0
requests==2.24.0